+++ extensions/default/src/DicomTagBrowser/DicomTagBrowser.tsx
@@ -2,7 +2,7 @@ import dcmjs from 'dcmjs';
 import moment from 'moment';
 import React, { useState, useMemo, useCallback } from 'react';
 import { classes, Types } from '@ohif/core';
-import { InputFilterText } from '@ohif/ui';
+import { InputFilter } from '@ohif/ui-next';
 import { Select, SelectTrigger, SelectContent, SelectItem, Slider } from '@ohif/ui-next';
 
 import DicomTagTable from './DicomTagTable';
@@ -18,7 +18,7 @@ export type Row = {
   depth: number;
   parents?: string[];
   children?: string[];
-  areChildrenVisible?: true;
+  areChildrenVisible?: boolean;
 };
 
 let rowCounter = 0;
@@ -91,7 +91,7 @@ const DicomTagBrowser = ({
 
     setShouldShowInstanceList(isImageStack && activeDisplaySet.images.length > 1);
     const tags = getSortedTags(metadata);
-    const rows = getFormattedRowsFromTags({ tags, metadata, depth: 0 });
+    const rows = getFormattedRowsFromTags({ tags, metadata });
     return rows;
   }, [getMetadata, activeDisplaySet]);
 
@@ -173,11 +173,17 @@ const DicomTagBrowser = ({
             <span className="text-muted-foreground flex h-6 items-center text-xs">
               Search metadata
             </span>
-            <InputFilterText
-              placeholder="Search metadata..."
-              onDebounceChange={setFilterValue}
-              className="text-foreground"
-            />
+            <InputFilter
+              className="text-muted-foreground"
+              onChange={setFilterValue}
+            >
+              <InputFilter.SearchIcon />
+              <InputFilter.Input
+                placeholder="Search metadata"
+                className="pl-9 pr-9"
+              />
+              <InputFilter.ClearButton />
+            </InputFilter>
           </div>
         </div>
       </div>
@@ -186,57 +192,106 @@ const DicomTagBrowser = ({
   );
 };
 
-function getFormattedRowsFromTags({ tags, metadata, depth, parents }) {
+function getFormattedRowsFromTags({ tags, metadata }) {
   const rows: Row[] = [];
+  const stack = [{ tags, depth: 0, parents: null, index: 0, children: [] }];
+  const parentChildMap = new Map();
 
-  tags.forEach(tagInfo => {
-    const uid = generateRowId();
-    if (tagInfo.vr === 'SQ') {
-      const children = tagInfo.values.flatMap(value =>
-        getFormattedRowsFromTags({
-          tags: value,
-          metadata,
-          depth: depth + 1,
-          parents: parents ? [...parents, uid] : [uid],
-        })
-      );
-      const row: Row = {
-        uid,
-        tag: tagInfo.tag,
-        valueRepresentation: tagInfo.vr,
-        keyword: tagInfo.keyword,
-        value: '',
-        depth,
-        isVisible: true,
-        areChildrenVisible: true,
-        children: children.map(child => child.uid),
-        parents,
-      };
-      rows.push(row, ...children);
-    } else {
-      if (tagInfo.vr === 'xs') {
-        try {
-          const tag = dcmjs.data.Tag.fromPString(tagInfo.tag).toCleanString();
-          const originalTagInfo = metadata[tag];
-          tagInfo.vr = originalTagInfo.vr;
-        } catch (error) {
-          console.warn(`Failed to parse value representation for tag '${tagInfo.keyword}'`);
+  while (stack.length > 0) {
+    const current = stack.pop();
+    const { tags, depth, parents, index, children } = current;
+
+    for (let i = index; i < tags.length; i++) {
+      const tagInfo = tags[i];
+      const uid = tagInfo.uid ?? generateRowId();
+
+      if (parents?.length > 0) {
+        parents.forEach(parent => {
+          parentChildMap.get(parent).push(uid);
+        });
+      }
+
+      if (tagInfo.vr === 'SQ') {
+        const row: Row = {
+          uid,
+          tag: tagInfo.tag,
+          valueRepresentation: tagInfo.vr,
+          keyword: tagInfo.keyword,
+          value: '',
+          depth,
+          isVisible: true,
+          areChildrenVisible: true,
+          children: [],
+          parents,
+        };
+        rows.push(row);
+        parentChildMap.set(uid, row.children);
+
+        const newParents = parents ? [...parents, uid] : [uid];
+
+        if (tagInfo.values.length > 0) {
+          stack.push({ tags, depth, parents, index: i + 1, children });
+          for (
+            let j = tagInfo.values.length - 1, values = tagInfo.values[j];
+            j >= 0;
+            values = tagInfo.values[--j]
+          ) {
+            const itemUid = generateRowId();
+            stack.push({
+              tags: values,
+              depth: depth + 2,
+              parents: [...newParents, itemUid],
+              index: 0,
+              children: [],
+            });
+            const itemTagInfo = {
+              tags: [
+                {
+                  tag: '(FFFE,E000)',
+                  vr: '',
+                  keyword: `Item #${j}`,
+                  value: '',
+                  uid: itemUid,
+                },
+              ],
+              depth: depth + 1,
+              parents: newParents,
+              index: 0,
+              children: [],
+            };
+            stack.push(itemTagInfo);
+            parentChildMap.set(itemUid, itemTagInfo.children);
+          }
+          break;
+        }
+      } else {
+        if (tagInfo.vr === 'xs') {
+          try {
+            const tag = dcmjs.data.Tag.fromPString(tagInfo.tag).toCleanString();
+            const originalTagInfo = metadata[tag];
+            tagInfo.vr = originalTagInfo.vr;
+          } catch (error) {
+            console.warn(`Failed to parse value representation for tag '${tagInfo.keyword}'`);
+          }
+        }
+        const row: Row = {
+          uid,
+          tag: tagInfo.tag,
+          valueRepresentation: tagInfo.vr,
+          keyword: tagInfo.keyword,
+          value: tagInfo.value,
+          depth,
+          isVisible: true,
+          parents,
+        };
+        rows.push(row);
+        if (row.tag === '(FFFE,E000)') {
+          row.areChildrenVisible = true;
+          row.children = [];
         }
       }
-      const row: Row = {
-        uid,
-        tag: tagInfo.tag,
-        valueRepresentation: tagInfo.vr,
-        keyword: tagInfo.keyword,
-        value: tagInfo.value,
-        depth,
-        isVisible: true,
-        parents,
-      };
-      rows.push(row);
     }
-  });
-
+  }
   return rows;
 }