+++ platform/ui-next/src/components/SegmentationTable/index.ts
@@ -3,20 +3,22 @@ import { SegmentationTableConfig } from './SegmentationTableConfig';
 import { AddSegmentationRow } from './AddSegmentationRow';
 import { AddSegmentRow } from './AddSegmentRow';
 import { SegmentationSegments } from './SegmentationSegments';
-import { SegmentationSelectorHeader } from './SegmentationSelectorHeader';
-import { SegmentationHeader } from './SegmentationHeader';
-import { useSegmentationTableContext } from './SegmentationTableContext';
+import {
+  useSegmentationTableContext,
+  useSegmentStatistics,
+  useSegmentationExpanded,
+} from './contexts';
 
 SegmentationTable.Segments = SegmentationSegments;
 SegmentationTable.Config = SegmentationTableConfig;
 SegmentationTable.AddSegmentRow = AddSegmentRow;
 SegmentationTable.AddSegmentationRow = AddSegmentationRow;
-SegmentationTable.SelectorHeader = SegmentationSelectorHeader;
-SegmentationTable.Header = SegmentationHeader;
 
 export {
   // context
   useSegmentationTableContext,
+  useSegmentStatistics,
+  useSegmentationExpanded,
   // components
   SegmentationTable,
 };